refactor: extract kv_cache_planning.py from kv_cache_utils.py
Move KV cache grouping/planning helpers out of `kv_cache_utils.py` into
a new `kv_cache_planning.py` module. The planning module now owns
"spec -> groups -> tensors": grouping strategies, page-size unification,
hybrid spec normalization, deepseek_v4 layout, per-worker projection, and
the top-level `get_kv_cache_configs` orchestrator. `kv_cache_utils.py`
keeps block hashing, free-block queue, request hashing, and the
single-spec memory-check helpers.

This is a pure move: function bodies, docstrings, and comments are
preserved verbatim. Only the import surface changes.

Callers updated:
- vllm/v1/engine/core.py: import `get_kv_cache_configs` from
  `kv_cache_planning`.
- vllm/v1/worker/gpu_model_runner.py: import
  `get_kv_cache_groups` / `get_kv_cache_config_from_groups` from
  `kv_cache_planning`.
- tests adjusted to follow the move.
